Split up the examples file.
